Setting up scaffolding for second attempt at top banner.
